refactor(Generators): 🎨 move computer data to user_agent module Make dependencis optional Seperate dev and live dependencies
