+++ Cargo.toml
@@ -47,13 +47,13 @@ beer = ["misc"]
 company = ["misc", "name"]
 colour = ["misc"]
 contact = ["misc", "internet"]
-currency = ["misc"]
+currency = ["misc", "libmath"]
 datetime =["misc", "chrono"]
 file =["misc"]
 generator =["misc", "person"]
 hacker = ["misc", "generator", "contact"]
 hipster = ["misc", "words"]
-image = []
+image = ["simplerand"]
 internet = ["misc", "person"]
 job = ["misc", "company"]
 language = ["misc"]
@@ -66,16 +66,19 @@ person = ["misc", "payment", "contact", "address"]
 random-bool = ["misc"]
 status-code = ["misc"]
 unique = ["misc", "chrono", "uuid"]
-user-agent = ["misc"]
+user-agent = ["misc", "datetime"]
 vehicle = ["misc"]
 words = ["misc", "company", "job", "name"]
 
 [dependencies]
-rand = "0.8.5"
-libmath = "0.2.1"
+libmath = { version = "0.2.1", optional = true}
 chrono = { version = "0.4", optional = true }
 uuid = { version = "1.0", features = ["v1", "v4"],  optional = true }
-simplerand = "1.3.0"
+simplerand = { version = "1.3.0", optional = true }
+
+[dev-dependencies]
+rand = "0.8.5"
+
 
 [package.metadata.docs.rs]
 all-features = true
